HTBHF-2677 Move TestConstants and shorted test data factory methods.
