+++ common_rest/src/test/java/uk/gov/dhsc/htbhf/eligibility/model/CombinedIdentityAndEligibilityResponseTest.java
@@ -3,45 +3,45 @@ package uk.gov.dhsc.htbhf.eligibility.model;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static uk.gov.dhsc.htbhf.eligibility.model.testhelper.CombinedIdentityAndEligibilityResponseTestDataFactory.anIdentityMatchFailedResponse;
-import static uk.gov.dhsc.htbhf.eligibility.model.testhelper.CombinedIdentityAndEligibilityResponseTestDataFactory.anIdentityMatchedEligibilityConfirmedUCResponseWithAllMatches;
-import static uk.gov.dhsc.htbhf.eligibility.model.testhelper.CombinedIdentityAndEligibilityResponseTestDataFactory.anIdentityMatchedEligibilityNotConfirmedResponse;
+import static uk.gov.dhsc.htbhf.eligibility.model.testhelper.CombinedIdAndEligibilityResponseTestDataFactory.anIdMatchFailedResponse;
+import static uk.gov.dhsc.htbhf.eligibility.model.testhelper.CombinedIdAndEligibilityResponseTestDataFactory.anIdMatchedEligibilityConfirmedUCResponseWithAllMatches;
+import static uk.gov.dhsc.htbhf.eligibility.model.testhelper.CombinedIdAndEligibilityResponseTestDataFactory.anIdMatchedEligibilityNotConfirmedResponse;
 
 class CombinedIdentityAndEligibilityResponseTest {
 
     @Test
     void shouldNotBeIneligibleForConfirmedEligibilityOutcome() {
-        CombinedIdentityAndEligibilityResponse response = anIdentityMatchedEligibilityConfirmedUCResponseWithAllMatches();
+        CombinedIdentityAndEligibilityResponse response = anIdMatchedEligibilityConfirmedUCResponseWithAllMatches();
         assertThat(response.isNotEligible()).isFalse();
     }
 
     @Test
     void shouldBeEligibleForConfirmedEligibilityOutcome() {
-        CombinedIdentityAndEligibilityResponse response = anIdentityMatchedEligibilityConfirmedUCResponseWithAllMatches();
+        CombinedIdentityAndEligibilityResponse response = anIdMatchedEligibilityConfirmedUCResponseWithAllMatches();
         assertThat(response.isEligible()).isTrue();
     }
 
     @Test
     void shouldBeIneligibleForNotSetEligibilityOutcome() {
-        CombinedIdentityAndEligibilityResponse response = anIdentityMatchFailedResponse();
+        CombinedIdentityAndEligibilityResponse response = anIdMatchFailedResponse();
         assertThat(response.isNotEligible()).isTrue();
     }
 
     @Test
     void shouldNotBeEligibleForNotSetEligibilityOutcome() {
-        CombinedIdentityAndEligibilityResponse response = anIdentityMatchFailedResponse();
+        CombinedIdentityAndEligibilityResponse response = anIdMatchFailedResponse();
         assertThat(response.isEligible()).isFalse();
     }
 
     @Test
     void shouldBeIneligibleForNotConfirmedEligibilityOutcome() {
-        CombinedIdentityAndEligibilityResponse response = anIdentityMatchedEligibilityNotConfirmedResponse();
+        CombinedIdentityAndEligibilityResponse response = anIdMatchedEligibilityNotConfirmedResponse();
         assertThat(response.isNotEligible()).isTrue();
     }
 
     @Test
     void shouldNotBeEligibleForNotConfirmedEligibilityOutcome() {
-        CombinedIdentityAndEligibilityResponse response = anIdentityMatchedEligibilityNotConfirmedResponse();
+        CombinedIdentityAndEligibilityResponse response = anIdMatchedEligibilityNotConfirmedResponse();
         assertThat(response.isEligible()).isFalse();
     }
